requirements.txt
openai>=0.27.0
langchain_openai>=0.0.1
langchain_community>=0.0.1
langchain_huggingface>=0.0.1
langchain_core>=0.0.1
langchain_text_splitter>=0.0.1
dateutil>=2.8.2
